feat(cfg): Refactor cfg implementation to use new IR implementation
