+++ src/lib.rs
@@ -1,6 +1,5 @@
 pub mod ast;
 pub mod cfg;
-pub mod instruction;
 pub mod ir;
 pub mod optim;
 pub mod parser;
